readme: Fix README formatting and typos
Current README has multiple formatting problems such as using images
instead of anchors for links and blockquotes for things that are not
quotes. It also have multiple typos in titles or texts.

I'm fixing these formatting problems and also changing the name of the
file from  `Readme.md` to `README.md` to better adhere to de-facto
standards.